UserViewModel

+++ app/src/main/java/com/example/culturunya/endpoints/events/Rating.kt
@@ -1,13 +1,13 @@
 package com.example.culturunya.endpoints.events
 
+import com.example.culturunya.endpoints.users.UserSimpleInfo
 import com.squareup.moshi.Json
-import java.util.Date
 
 data class Rating (
     @Json(name="user")
     val user: UserSimpleInfo,
     @Json(name = "id")
-    val id: Int,
+    val id: Long,
     @Json(name="date")
     val date: String,
     @Json(name="rating")
